# docs/api/dcrs-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DCRS API for the Mitra mobile app (the Audit Assistant)",
    "version": "1.3.0",
    "summary": "Read and change DCRS data as a signed-in person, through DCRS's own rules.",
...
  },
  "servers": [
    {
      "url": "http://{host}:{port}",
      "description": "The DCRS server on the factory network (API_PORT, 4000 unless set).",
      "variables": {
        "host": {
          "default": "localhost",
          "description": "The DCRS computer's name or address."
        },
        "port": {
          "default": "4000"
        }
      }
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/api/notifications": {
      "get": {
        "tags": [
          "Notifications",
          "Read"
        ],
        "operationId": "webNotifications",
        "summary": "READ - the person's notifications",
        "description": "The signed-in person's notifications (REQUIREMENTS §97), each worded in the language asked. Only the person's own: nobody reads another's. The website's own route, by its session cookie (the bell and the Notifications page); the phone uses the /api/v1 twin.",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "description": "open: the items still needing the person; all: every item of the last 60 days.",
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "all"
              ],
              "default": "open"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "description": "The id of the last item already shown: the next page is the items older than it.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "description": "The language the titles and bodies are worded in.",
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "hi",
                "gu"
              ],
              "default": "en"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The person's notifications, newest first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationList"
                },
                "example": {
                  "items": [
                    {
                      "id": 412,
                      "kind": "needs_input",
                      "priority": "high",
                      "title": "12 readings to enter: Daily Pest Control Monitoring Record",
                      "body": "F/HR/17 Daily Pest Control Monitoring Record of 09-Oct-2026 is ready for you: 12 readings to enter, then submit.",
                      "data": {
                        "documentId": "daily-pest-monitoring",
                        "formatNo": "F/HR/17",
                        "documentName": "Daily Pest Control Monitoring Record",
                        "module": "HR",
                        "recordId": "rec-mgj2k1-7-abcd12",
                        "dueDate": "2026-10-09",
                        "count": 12
                      },
                      "createdAt": "2026-10-09T03:00:12.000Z",
                      "readAt": null,
                      "resolvedAt": null
                    }
                  ],
                  "unread": 1,
                  "open": 1
                }
              }
            }
          },
          "400": {
            "description": "A value this route does not accept.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "bad-request": {
                    "value": {
                      "error": "lang must be en, hi or gu.",
                      "code": "bad-request"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NotSignedIn"
          },
          "403": {
            "$ref": "#/components/responses/PasswordChangeRequired"
          }
        }
      }
    },
    "/api/notifications/read": {
      "post": {
        "tags": [
          "Notifications",
          "Change"
        ],
        "operationId": "webNotificationsRead",
        "summary": "CHANGE - mark notifications read",
        "description": "Marks the person's own notifications read: the ones named, or every one. The website's own route, by its session cookie (the bell and the Notifications page); the phone uses the /api/v1 twin.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadRequest"
              },
              "examples": {
                "some": {
                  "value": {
                    "ids": [
                      412,
                      409
                    ]
                  }
                },
                "all": {
                  "value": {
                    "all": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Marked; the unread count now.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadResponse"
                },
                "example": {
                  "unread": 0
                }
              }
            }
          },
          "400": {
            "description": "A value this route does not accept.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "bad-request": {
                    "value": {
                      "error": "Send ids (a list of at most 500 numbers) or all: true.",
                      "code": "bad-request"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NotSignedIn"
          },
          "403": {
            "$ref": "#/components/responses/PasswordChangeRequired"
          }
        }
      }
    },
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "Notifications",
          "Mitra",
          "Read"
        ],
        "operationId": "notifications",
        "summary": "READ - the person's notifications",
        "description": "The signed-in person's notifications (REQUIREMENTS §97): the phone's inbox, each worded in the language asked. Only the person's own: nobody reads another's. The super admin's include the morning and evening summaries and the escalations.",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "description": "open: the items still needing the person; all: every item of the last 60 days.",
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "all"
              ],
              "default": "open"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "description": "The id of the last item already shown: the next page is the items older than it.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "description": "The language the titles and bodies are worded in.",
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "hi",
                "gu"
              ],
              "default": "en"
            }
          },
          {
            "$ref": "#/components/parameters/ClientName"
          }
        ],
        "responses": {
          "200": {
            "description": "The person's notifications, newest first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationList"
                },
                "example": {
                  "items": [
                    {
                      "id": 412,
                      "kind": "needs_input",
                      "priority": "high",
                      "title": "12 readings to enter: Daily Pest Control Monitoring Record",
                      "body": "F/HR/17 Daily Pest Control Monitoring Record of 09-Oct-2026 is ready for you: 12 readings to enter, then submit.",
                      "data": {
                        "documentId": "daily-pest-monitoring",
                        "formatNo": "F/HR/17",
                        "documentName": "Daily Pest Control Monitoring Record",
                        "module": "HR",
                        "recordId": "rec-mgj2k1-7-abcd12",
                        "dueDate": "2026-10-09",
                        "count": 12
                      },
                      "createdAt": "2026-10-09T03:00:12.000Z",
                      "readAt": null,
                      "resolvedAt": null
                    }
                  ],
                  "unread": 1,
                  "open": 1
                }
              }
            }
          },
          "400": {
            "description": "A value this route does not accept.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "bad-request": {
                    "value": {
                      "error": "lang must be en, hi or gu.",
                      "code": "bad-request"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NotSignedIn"
          },
          "403": {
            "$ref": "#/components/responses/PasswordChangeRequired"
          }
        }
      }
    },
    "/api/v1/notifications/read": {
      "post": {
        "tags": [
          "Notifications",
          "Mitra",
          "Change"
        ],
        "operationId": "notificationsRead",
        "summary": "CHANGE - mark notifications read",
        "description": "Marks the person's own notifications read: the ones named, or every one (\"Mark all read\").",
        "parameters": [
          {
            "$ref": "#/components/parameters/ClientName"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadRequest"
              },
              "examples": {
                "some": {
                  "value": {
                    "ids": [
                      412,
                      409
                    ]
                  }
                },
                "all": {
                  "value": {
                    "all": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Marked; the unread count now.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadResponse"
                },
                "example": {
                  "unread": 0
                }
              }
            }
          },
          "400": {
            "description": "A value this route does not accept.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "bad-request": {
                    "value": {
                      "error": "Send ids (a list of at most 500 numbers) or all: true.",
                      "code": "bad-request"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NotSignedIn"
          },
          "403": {
            "$ref": "#/components/responses/PasswordChangeRequired"
          }
        }
      }
    },
    "/api/v1/notifications/test": {
      "post": {
        "tags": [
          "Notifications",
          "Mitra",
          "Change"
        ],
        "operationId": "notificationsTest",
        "summary": "CHANGE - send a test push to the caller's own phones",
        "description": "Sends \"This is a test notification from DCRS\" to every phone the caller registered, in each phone's language. Sends nothing when push is switched off on the server (PUSH_ENABLED=0) or no phone is registered, and says why. At most one test every 20 seconds per person.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ClientName"
          }
        ],
        "responses": {
          "200": {
            "description": "How many phones it was sent to.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestPushResponse"
                },
                "examples": {
                  "sent": {
                    "value": {
                      "sent": 1
                    }
                  },
                  "off": {
                    "value": {
                      "sent": 0,
                      "reason": "No phone is registered for notifications on this account yet."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NotSignedIn"
          },
          "403": {
            "$ref": "#/components/responses/PasswordChangeRequired"
          },
          "429": {
            "description": "A test was sent a moment ago.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "too-many": {
                    "value": {
                      "error": "A test was sent a moment ago. Wait 20 seconds and try again.",
                      "code": "too-many"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notification-preferences": {
      "get": {
        "tags": [
          "Notifications",
          "Mitra",
          "Read"
        ],
        "operationId": "notificationPreferences",
        "summary": "READ - what is pushed to this person",
        "description": "Every kind with whether it is pushed (true unless the person switched it off), and the phone's own reminders when the person has chosen.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ClientName"
          }
        ],
        "responses": {
          "200": {
            "description": "The person's choices.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferences"
                },
                "example": {
                  "kinds": {
                    "ready": true,
                    "needs_input": true,
                    "due": true,
                    "upcoming": false,
                    "overdue": true,
                    "verify": true,
                    "sent_back": true,
                    "boss_summary": true,
                    "escalation": true,
                    "access_changed": true
                  },
                  "reminders": false
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NotSignedIn"
          },
          "403": {
            "$ref": "#/components/responses/PasswordChangeRequired"
          }
        }
      },
      "put": {
        "tags": [
          "Notifications",
          "Mitra",
          "Change"
        ],
        "operationId": "notificationPreferencesSet",
        "summary": "CHANGE - choose what is pushed",
        "description": "A kind set to false is not pushed; it still reaches the inbox. Kinds left out keep their setting. Answers every kind, as GET does.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ClientName"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationPreferences"
              },
              "example": {
                "kinds": {
                  "upcoming": false
                },
                "reminders": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The person's choices now.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferences"
                },
                "example": {
                  "kinds": {
                    "ready": true,
                    "needs_input": true,
                    "due": true,
                    "upcoming": false,
                    "overdue": true,
                    "verify": true,
                    "sent_back": true,
                    "boss_summary": true,
                    "escalation": true,
                    "access_changed": true
                  },
                  "reminders": false
                }
              }
            }
          },
          "400": {
            "description": "A value this route does not accept.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "bad-request": {
                    "value": {
                      "error": "kinds must name notification kinds, each true or false.",
                      "code": "bad-request"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NotSignedIn"
          },
          "403": {
            "$ref": "#/components/responses/PasswordChangeRequired"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "The dcrs_session cookie's value from POST /api/auth/login. Valid until the close of the day it was started (the sign-in answer's session.endsAt)."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dcrs_session"
      }
    },
...
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Every refusal. `error` is a sentence to show a person; `code` is for the program. A refusal outside the staff's working hours also says when they start again (`opensAt`).",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "opensAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Only with the code `outside-working-hours`: the next moment DCRS opens (the start of the next working day), or null when the plant's calendar has no working day in the year ahead."
          }
        }
      },
...
      "NotificationList": {
        "type": "object",
        "required": [
          "items",
          "unread",
          "open"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationItem"
            },
            "description": "Newest first."
          },
          "unread": {
            "type": "integer",
            "description": "Open items not yet read (the bell's count)."
          },
          "open": {
            "type": "integer",
            "description": "Open items, read or not (the app's badge)."
          }
        }
      },
      "ReadRequest": {
        "type": "object",
        "description": "Either ids (at most 500) or all: true.",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "maxItems": 500
          },
          "all": {
            "type": "boolean",
            "const": true
          }
        }
      },
...
    }
  }
}